Cleanup result assertion in ESQL tests
