+++ x-pack/plugin/esql/qa/server/src/main/java/org/elasticsearch/xpack/esql/qa/rest/EsqlSpecTestCase.java
@@ -17,21 +17,13 @@ import org.elasticsearch.client.ResponseException;
 import org.elasticsearch.client.RestClient;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.xcontent.XContentHelper;
-import org.elasticsearch.core.Types;
-import org.elasticsearch.exponentialhistogram.ExponentialHistogramXContent;
 import org.elasticsearch.features.NodeFeature;
-import org.elasticsearch.geometry.Geometry;
-import org.elasticsearch.geometry.Point;
-import org.elasticsearch.geometry.utils.GeometryValidator;
-import org.elasticsearch.geometry.utils.WellKnownText;
 import org.elasticsearch.index.mapper.MappedFieldType;
 import org.elasticsearch.logging.LogManager;
 import org.elasticsearch.logging.Logger;
 import org.elasticsearch.test.MapMatcher;
 import org.elasticsearch.test.rest.ESRestTestCase;
 import org.elasticsearch.test.rest.TestFeatureService;
-import org.elasticsearch.xcontent.XContentParser;
-import org.elasticsearch.xcontent.XContentParserConfiguration;
 import org.elasticsearch.xcontent.XContentType;
 import org.elasticsearch.xpack.esql.CsvSpecReader.CsvTestCase;
 import org.elasticsearch.xpack.esql.CsvTestUtils;
@@ -51,9 +43,6 @@ import org.junit.Before;
 import org.junit.Rule;
 
 import java.io.IOException;
-import java.math.BigDecimal;
-import java.math.MathContext;
-import java.math.RoundingMode;
 import java.net.URL;
 import java.util.Arrays;
 import java.util.List;
@@ -66,13 +55,9 @@ import java.util.stream.IntStream;
 import java.util.stream.LongStream;
 import java.util.stream.Stream;
 
-import static org.apache.lucene.geo.GeoEncodingUtils.decodeLatitude;
-import static org.apache.lucene.geo.GeoEncodingUtils.decodeLongitude;
-import static org.apache.lucene.geo.GeoEncodingUtils.encodeLatitude;
-import static org.apache.lucene.geo.GeoEncodingUtils.encodeLongitude;
 import static org.elasticsearch.test.MapMatcher.assertMap;
 import static org.elasticsearch.test.MapMatcher.matchesMap;
-import static org.elasticsearch.xpack.esql.CsvAssert.assertData;
+import static org.elasticsearch.xpack.esql.CsvAssert.assertDataWithValueConverter;
 import static org.elasticsearch.xpack.esql.CsvAssert.assertMetadata;
 import static org.elasticsearch.xpack.esql.CsvSpecReader.specParser;
 import static org.elasticsearch.xpack.esql.CsvTestUtils.ExpectedResults;
@@ -500,74 +485,17 @@ public abstract class EsqlSpecTestCase extends ESRestTestCase {
         List<List<Object>> actualValues,
         Logger logger
     ) {
-        assertMetadata(expected, actualColumns, logger);
-        assertData(expected, actualValues, testCase.ignoreOrder, ignoreValueOrder(), logger, this::valueMapper);
-    }
-
-    private Object valueMapper(CsvTestUtils.Type type, Object value) {
-        if (value == null) {
-            return "null";
-        }
-        if (value instanceof CsvTestUtils.Range) {
-            return value;
-        }
-        if (type == CsvTestUtils.Type.GEO_POINT || type == CsvTestUtils.Type.CARTESIAN_POINT) {
-            // Point tests are failing in clustered integration tests because of tiny precision differences at very small scales
-            if (value instanceof String wkt) {
-                try {
-                    Geometry geometry = WellKnownText.fromWKT(GeometryValidator.NOOP, false, wkt);
-                    if (geometry instanceof Point point) {
-                        return normalizedPoint(type, point.getX(), point.getY());
-                    }
-                } catch (Throwable ignored) {}
-            }
-        }
-        if (type == CsvTestUtils.Type.EXPONENTIAL_HISTOGRAM) {
-            if (value instanceof Map<?, ?> map) {
-                return ExponentialHistogramXContent.parseForTesting(Types.<Map<String, Object>>forciblyCast(map));
-            }
-            if (value instanceof String json) {
-                try (XContentParser parser = XContentType.JSON.xContent().createParser(XContentParserConfiguration.EMPTY, json)) {
-                    return ExponentialHistogramXContent.parseForTesting(parser);
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
-                }
-            }
-        }
-        if (value instanceof List<?> vs) {
-            return vs.stream().map(v -> valueMapper(type, v)).toList();
-        }
-        if (type == CsvTestUtils.Type.DOUBLE && enableRoundingDoubleValuesOnAsserting()) {
-            if (value instanceof Double d) {
-                if (Double.isNaN(d) || Double.isInfinite(d)) {
-                    return d;
-                }
-                return new BigDecimal(d).round(new MathContext(7, RoundingMode.HALF_DOWN)).doubleValue();
-            } else if (value instanceof String s) {
-                if ("NaN".equals(s)) {
-                    return Double.NaN;
-                }
-                return new BigDecimal(s).round(new MathContext(7, RoundingMode.HALF_DOWN)).doubleValue();
-            }
-        }
-        if (type == CsvTestUtils.Type.TEXT || type == CsvTestUtils.Type.KEYWORD || type == CsvTestUtils.Type.SEMANTIC_TEXT) {
-            if (value instanceof String s) {
-                value = s.replaceAll("\\\\n", "\n");
-            }
-        }
-        if (type == CsvTestUtils.Type.DOUBLE) {
-            if (value instanceof String s && "NaN".equals(s)) {
-                return Double.NaN;
-            }
-            return ((Number) value).doubleValue();
-        }
-        if (type == CsvTestUtils.Type.INTEGER) {
-            return ((Number) value).intValue();
-        }
-        if (type == CsvTestUtils.Type.LONG) {
-            return ((Number) value).longValue();
-        }
-        return value.toString();
+        var actualColumnNames = actualColumns.stream().map(c -> c.get("name")).toList();
+        var actualColumnTypes = actualColumns.stream().map(c -> CsvTestUtils.Type.asType(c.get("type"))).toList();
+        assertMetadata(expected, actualColumnNames, actualColumnTypes, List.of(), logger);
+        assertDataWithValueConverter(
+            expected,
+            actualValues,
+            testCase.ignoreOrder,
+            ignoreValueOrder(),
+            enableRoundingDoubleValuesOnAsserting(),
+            logger
+        );
     }
 
     /**
@@ -576,19 +504,6 @@ public abstract class EsqlSpecTestCase extends ESRestTestCase {
      */
     protected boolean enableRoundingDoubleValuesOnAsserting() {
         return false;
-    }
-
-    private static String normalizedPoint(CsvTestUtils.Type type, double x, double y) {
-        if (type == CsvTestUtils.Type.GEO_POINT) {
-            return normalizedGeoPoint(x, y);
-        }
-        return String.format(Locale.ROOT, "POINT (%f %f)", (float) x, (float) y);
-    }
-
-    private static String normalizedGeoPoint(double x, double y) {
-        x = decodeLongitude(encodeLongitude(x));
-        y = decodeLatitude(encodeLatitude(y));
-        return String.format(Locale.ROOT, "POINT (%f %f)", x, y);
     }
 
     private Throwable reworkException(Throwable th) {
